requirements.txt
Flask==0.10.1
Flask-Cache==0.13.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
gunicorn==19.3.0
itsdangerous==0.24
#pylibmc==1.4.2
python-postmark==0.4.7
wsgiref==0.1.2
